+++ src/utils/downloader.c
@@ -3160,7 +3160,7 @@ GF_Err gf_dm_get_file_memory(const char *url, char **out_data, u32 *out_size, ch
 		e = gf_cache_close_write_cache(dnload->cache_entry, dnload, e == GF_OK);
 
 	if (!e) {
-		u32 size = ftell(f), read;
+		u32 size = (u32) ftell(f), read;
 		*out_size = size;
 		*out_data = gf_malloc(sizeof(char)* ( 1 + size));
 		fseek(f, 0, SEEK_SET);
